% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_game_latest.R
\name{calc_game_latest}
\alias{calc_game_latest}
\title{Find the index of the latest game so far in this Premier League season}
\usage{
calc_game_latest(results)
}
\arguments{
\item{results}{These are the results generated by running \code{get_results}.}
}
\description{
This function finds the index of the most recently played game in this Premier League season.
}
\examples{
\dontrun{
calc_game_latest(
  results = data_results
  )
}

}
\keyword{Sport}
