% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_points_expected_remaining.R
\name{calc_points_expected_remaining}
\alias{calc_points_expected_remaining}
\title{Calculate the points expected to be gained by each team in the remainder of this Premier League season}
\usage{
calc_points_expected_remaining(games_remaining)
}
\arguments{
\item{games_remaining}{These are the remaining games, with their associated model parameters, as generated by running \code{model_parameters_unplayed}.}
}
\description{
This function calculates the points expected to be gained by each team in the remainder of this Premier League season.
}
\examples{
\dontrun{
calc_points_expected_remaining(
  games_remaining = data_results
  )
}

}
\keyword{Sport}
