% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_points_expected_total.R
\name{calc_points_expected_total}
\alias{calc_points_expected_total}
\title{Calculate the points expected to be gained by each team across this Premier League season}
\usage{
calc_points_expected_total(table_current, points_expected)
}
\arguments{
\item{table_current}{This is the current table, as generated by running \code{calc_table_current}.}

\item{points_expected}{These are the expected points per team across the rest of the season, as generated by running \code{calc_points_expected_remaining}.}
}
\description{
This function projects the points expected to be gained by each team across this Premier League season.
}
\examples{
\dontrun{
calc_points_expected_total(
  table_current = data_table_current,
  points_expected = data_points_expected_remaining
  )
}

}
\keyword{Sport}
