% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_result_latest.R
\name{calc_result_latest}
\alias{calc_result_latest}
\title{Get the latest result available from this Premier League season}
\usage{
calc_result_latest(results)
}
\arguments{
\item{results}{These are the results generated by running \code{get_results}.}
}
\description{
This function takes the latest result available from this Premier League season.
}
\examples{
\dontrun{
calc_result_latest(
  results = data_results
  )
}

}
\keyword{Sport}
