% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_table_current.R
\name{calc_table_current}
\alias{calc_table_current}
\title{Calculate the current table for this Premier League season}
\usage{
calc_table_current(results)
}
\arguments{
\item{results}{These are the results generated by running \code{get_results_filtered}.}
}
\description{
This function generates the latest standings of this Premier League season.
}
\examples{
\dontrun{
calc_table_current(
  results = data_results
  )
}

}
\keyword{Sport}
