% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_footballData.R
\name{get_footballData}
\alias{get_footballData}
\title{Get the latest available results for the Premier League in a given season}
\source{
\url{https://www.football-data.co.uk}
}
\usage{
get_footballData(value_link, table_schedule, table_teams, value_yearEnd)
}
\arguments{
\item{value_link}{This is the link for the data on the web. For example, you could use 'https://www.football-data.co.uk/mmz4281/2526/E0.csv'.}

\item{table_schedule}{This is the location of the schedule data, as generated through an in-built dataset or by using \code{get_openData_schedule}.}

\item{table_teams}{These are the teams in the season's Premier League, available as the \code{teams} dataset in this package.}

\item{value_yearEnd}{This is the integer required as the year in which the season ends.}
}
\description{
This function retrieves the latest data on the Premier League results for a given season.
}
\examples{
\dontrun{
get_footballData(
  value_link = "https://www.football-data.co.uk/mmz4281/2526/E0.csv",
  table_schedule = schedule_thisSeason,
  table_teams = teams,
  value_yearEnd = 2026L
  )
}

}
\keyword{Sport}
