% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_openData.R
\name{get_openData}
\alias{get_openData}
\title{Get the latest available results for the Premier League in a given season}
\source{
\url{https://github.com/openfootball/football.json}
}
\usage{
get_openData(value_path, table_teams, value_yearEnd)
}
\arguments{
\item{value_path}{This  is the location of the data on GitHub. See the example below for reference and use an address of the form, 'https://raw.githubusercontent.com/openfootball/football.json/refs/heads/master/2024-25/en.1.json'.}

\item{table_teams}{These are the teams in the season's Premier League, available as the \code{teams} dataset in this package.}

\item{value_yearEnd}{This is the integer required as the year in which the season ends.}
}
\description{
This function retrieves the latest data on the Premier League results for a given season.
}
\examples{
\dontrun{
get_openData(
  value_path = "https://raw.githubusercontent.com/openfootball/football.json/refs/etc",
  table_teams = teams,
  value_yearEnd = 2025L
  )
}

}
\keyword{Sport}
