% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results.R
\name{get_results}
\alias{get_results}
\title{Get the Premier League results for the desired seasons}
\usage{
get_results(results_thisSeason, seasons = 0L)
}
\arguments{
\item{results_thisSeason}{These are the results generated by running \code{get_openData} on the current season, such as at 'https://raw.githubusercontent.com/openfootball/football.json/refs/heads/master/2024-25/en.1.json'.}

\item{seasons}{This is the integer required for the number of previous seasons to include. It defaults to zero.}
}
\description{
This function takes the latest data on this Premier League season and combines it with corresponding results from previous seasons, if desired.
}
\examples{
\dontrun{
get_results(
  results_thisSeason = data_thisSeason,
  seasons = 1L
  )
}

}
\keyword{Sport}
