% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_results_filtered.R
\name{get_results_filtered}
\alias{get_results_filtered}
\title{Get the Premier League results for the desired seasons}
\usage{
get_results_filtered(results, index_game_latest, lookback_rounds)
}
\arguments{
\item{results}{These are the results from this and possibly earlier seasons, as generated by \code{get_results}.}

\item{index_game_latest}{This is the index of the latest game played, which can be generated by \code{calc_game_latest}.}

\item{lookback_rounds}{This is the number of rounds of fixtures to use in a model (so 38 would represent a whole season).}
}
\description{
This function takes the latest data on this Premier League season and combines it with corresponding results from previous seasons, if desired.
}
\examples{
\dontrun{
get_results_filtered(
  results = data_results,
  index_game_latest = 280L,
  lookback_rounds = 38L,
  )
}

}
\keyword{Sport}
