% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_extract_parameters.R
\name{model_extract_parameters}
\alias{model_extract_parameters}
\title{Run the prediction model}
\usage{
model_extract_parameters(model_output)
}
\arguments{
\item{model_output}{This is the output generated by \code{model_run}.}
}
\description{
This function takes the relevant results generated in modelling the strength of the Premier League teams and extracts the relevant parameters.
}
\examples{
\dontrun{
model_extract_parameters(
  model_output = data_model_output
  )
}

}
\keyword{Sport}
