% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_parameters_unplayed.R
\name{model_parameters_unplayed}
\alias{model_parameters_unplayed}
\title{Assign the model parameters for unplayed games in this Premier League season.}
\usage{
model_parameters_unplayed(results, model_parameters)
}
\arguments{
\item{results}{These are the results from this and possibly earlier seasons, as generated by \code{get_results_filtered}.}

\item{model_parameters}{This is the output generated by \code{model_extract_parameters}.}
}
\description{
This function takes the relevant parameters from the model and assigns them to the approapriate teams in each remaining game of the season.
}
\examples{
\dontrun{
model_parameters_unplayed(
  model_parameters = data_model_parameters,
  results = data_results
  )
}

}
\keyword{Sport}
