% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_prepare_frame.R
\name{model_prepare_frame}
\alias{model_prepare_frame}
\title{Prepare the modelframe in order to run the prediction model}
\usage{
model_prepare_frame(results)
}
\arguments{
\item{results}{These are the results from this and possibly earlier seasons, as generated by \code{get_results_filtered}.}
}
\description{
This function takes the relevant filtered results from the Premier League and combines it in a way that the R prediction model can recognise.
}
\examples{
\dontrun{
model_prepare_frame(
  results = data_results
  )
}

}
\keyword{Sport}
