% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run.R
\name{model_run}
\alias{model_run}
\title{Run the prediction model}
\usage{
model_run(modelframe)
}
\arguments{
\item{modelframe}{This is the modelframe generated in \code{model_prepare_frame}.}
}
\description{
This function takes the relevant filtered results from the Premier League and uses them to model each team's capabilities, both at home and away.
}
\examples{
\dontrun{
model_run(
  modelframe = data_modelframe
  )
}

}
\keyword{Sport}
