% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremPredict-package.R
\docType{package}
\name{PremPredict-package}
\alias{PremPredict}
\alias{PremPredict-package}
\title{PremPredict: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p0bs/PremPredict}
  \item Report bugs at \url{https://github.com/p0bs/PremPredict/issues}
}

}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
\keyword{internal}
