% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_table_current.R
\name{print_table_current}
\alias{print_table_current}
\title{Print the current table for this Premier League season}
\usage{
print_table_current(table_latest)
}
\arguments{
\item{table_latest}{These are the latest standings generated by running \code{calc_table_current}.}
}
\description{
This function prints the latest standings of this Premier League season.
}
\examples{
\dontrun{
print_table_current(
  table_latest = data_calc_table_current
  )
}

}
\keyword{Sport}
