% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reformat_outcomes.R
\name{reformat_outcomes}
\alias{reformat_outcomes}
\title{Reformat the outcomes data for improved presentation.}
\usage{
reformat_outcomes(value)
}
\arguments{
\item{value}{This is the outcome value to be reformatted.}
}
\description{
This function takes the likelihoods of all possible standings for all clubs over this Premier League season and reformats them for improved presentation..
}
\examples{
\dontrun{
reformat_outcomes(
  value = 0.94
  )
}
}
\keyword{Sport}
