% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Assign the model parameters for unplayed games in this Premier League season.}
\usage{
run_simulations(
  results_thisSeason,
  number_seasons = 0L,
  lookback_rounds = 19L,
  number_simulations = 25000L,
  value_seed = 120519L
)
}
\arguments{
\item{results_thisSeason}{These are the results generated by running \code{get_openData} on the current season, such as at 'https://raw.githubusercontent.com/openfootball/football.json/refs/heads/master/2024-25/en.1.json'. The data should have the following columns and names:
\describe{
\item{number_match}{The id values of the matches in the dataset as a character, typically from "001"}
\item{number_match_integer}{the integer equivalent of the characters in \code{number_match}}
\item{matchday}{The date of the match in yyyy-mm-dd format}
\item{homeTeam}{The \code{shortName} of the home team, consistent with the data in the \code{teams} table}
\item{awayTeam}{The \code{shortName} of the away team, consistent with the data in the \code{teams} table}
\item{FTHG}{The integer number of goals scored in the whole match by the home team}
\item{FTAG}{The integer number of goals scored in the whole match by the away team}
\item{FTR}{The result of the match, as a factor of three levels ... where 'H', 'D' and 'A' represent a home win, a draw and an away win, respectively}
\item{played}{A logical to show if the game has yet been played}
\item{year_end}{The four figure integer value of the year in which the season ends}
}}

\item{number_seasons}{This is the integer required for the number of previous seasons to include. It defaults to zero.}

\item{lookback_rounds}{This is the integer number of rounds of fixtures to use in a model (so 38L would represent a whole season). Defaults to half a season (that is, 19L).}

\item{number_simulations}{This is the integer number of simulations to use for each game in the remaining season. Defaults to 25000L.}

\item{value_seed}{This is the integer seed to use for the random numbers in the simulation. Defaults to 120519L (which, IMHO, was a great footballing day).}
}
\description{
This function takes the relevant parameters from the model and assigns them to the approapriate teams in each remaining game of the season.
}
\examples{
\dontrun{
run_simulations(
  results_thisSeason = example_thisSeason,
  number_seasons = 1L,
  lookback_rounds = 78L,
  number_simulations = 25000L,
  value_seed = 120519L
  )
  }

}
\keyword{Sport}
