% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_games.R
\name{simulate_games}
\alias{simulate_games}
\title{Assign the model parameters for unplayed games in this Premier League season.}
\usage{
simulate_games(
  data_model_parameters_unplayed,
  value_number_sims = 50000,
  value_seed = 120519L
)
}
\arguments{
\item{data_model_parameters_unplayed}{These are the model parameters assigned to unplayed games in this Premier League season, as generated by \code{model_parameters_unplayed}.}

\item{value_number_sims}{This is the number of simulations to use for each game in the remaining season. Defaults to 50,000.}

\item{value_seed}{This is the seed to use for the random numbers in the simulation. Defaults to 120519 (which, IMHO, was a great footballing day).}
}
\description{
This function takes the relevant parameters from the model and assigns them to the approapriate teams in each remaining game of the season.
}
\examples{
\dontrun{
simulate_games(
  data_model_parameters_unplayed = data_model_parameters_unplayed,
  value_number_sims = 1000000,
  value_seed = 120519L
  )
}

}
\keyword{Sport}
