% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_outcomes.R
\name{simulate_outcomes}
\alias{simulate_outcomes}
\title{Show the likelihoods of all possible standings for all clubs over this Premier League season.}
\usage{
simulate_outcomes(data_standings_simulations, value_number_sims)
}
\arguments{
\item{data_standings_simulations}{These are the standings for each club in each scenario run, as generated by \code{simulate_standings}.}

\item{value_number_sims}{This is the number of simulations to use for each game in the remaining season. Defaults to 50,000.}
}
\description{
This function takes the relevant data from the simulations and the current table to the likelihoods of all possible standings for all clubs over this Premier League season.
}
\examples{
\dontrun{
simulate_outcomes(
  data_standings_simulations = data_standings_simulations,
  value_number_sims = 1000000
  )
}

}
\keyword{Sport}
