% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_standings.R
\name{simulate_standings}
\alias{simulate_standings}
\title{Simulate the standings over this Premier League season.}
\usage{
simulate_standings(data_game_simulations, data_table_latest)
}
\arguments{
\item{data_game_simulations}{These are the outcome scenarios of the unplayed games in this Premier League season, as generated by \code{simulate_games}.}

\item{data_table_latest}{These are the latest standings generated by running \code{calc_table_current}.}
}
\description{
This function takes the relevant data from the simulations and finds the standings for each team at the end of the season.
}
\examples{
\dontrun{
simulate_standings(
  data_game_simulations = data_game_simulations,
  data_table_latest = data_table_latest
  )
}

}
\keyword{Sport}
