% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-teams.R
\docType{data}
\name{teams}
\alias{teams}
\title{A dataset containing the teams in the latest Premier League season.}
\format{
A data frame with 20 rows (one for each team) and 4 variables:
\describe{
\item{teamName}{The long name of the team in question, used in footballData}
\item{shortName}{The three-letter code of the team in question}
\item{midName}{The more-readable name of the team in question}
\item{openName}{The more-readable name of the team in question, used in openData}
}
}
\source{
\url{https://www.premierleague.com/}
}
\usage{
teams
}
\description{
A dataset containing the teams in the latest Premier League season.
}
\keyword{datasets}
