% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_points_simulated_match.R
\name{calc_points_simulated_match}
\alias{calc_points_simulated_match}
\title{Calculate the result of one match, given model parameters and a uniform random variate.}
\usage{
calc_points_simulated_match(
  data_model_parameters_unplayed_slim,
  randoms,
  number_sims,
  value_match
)
}
\arguments{
\item{data_model_parameters_unplayed_slim}{These are the model parameters assigned to unplayed matches in this Premier League season, as generated by \code{model_parameters_unplayed}, with appropriate columns retained.}

\item{randoms}{This is a vector of uniformly-distributed random numbers, with a length of the product of the number of remaining matches and the number of simulations.}

\item{number_sims}{This is the number of simulations to use for each game in the remaining season. Defaults to 50,000.}

\item{value_match}{This is the simulated match to choose within the \code{randoms} vector of many simulated matches.}
}
\description{
This function takes the relevant outcome probabilities for a match and calculates its projected result based upon the outcome of a random number generator.
}
\examples{
\dontrun{
calc_points_simulated_match(
  data_model_parameters_unplayed_slim = data_model_parameters_unplayed,
  randoms = data_randoms,
  number_sims = value_number_sims,
  value_match = 1L
  )
}

}
\keyword{Sport}
