% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PremPredict-package.R
\docType{package}
\name{PremPredict-package}
\alias{PremPredict}
\alias{PremPredict-package}
\title{PremPredict: Predict the Premier League}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides functions to predict the outcome of Premier League games and seasons.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/p0bs/PremPredict}
  \item \url{https://p0bs.github.io/PremPredict/}
  \item Report bugs at \url{https://github.com/p0bs/PremPredict/issues}
}

}
\author{
\strong{Maintainer}: Robin Penfold \email{robin.penfold@icloud.com}

}
\keyword{internal}
