% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simulations.R
\name{run_simulations}
\alias{run_simulations}
\title{Assign the model parameters for unplayed games in this Premier League season.}
\usage{
run_simulations(
  results_thisSeason,
  number_seasons,
  lookback_rounds,
  number_simulations,
  value_seed
)
}
\arguments{
\item{results_thisSeason}{These are the results generated by running \code{get_openData} on the current season, such as at 'https://raw.githubusercontent.com/openfootball/football.json/refs/heads/master/2024-25/en.1.json'.}

\item{number_seasons}{This is the integer required for the number of previous seasons to include. It defaults to zero.}

\item{lookback_rounds}{This is the number of rounds of fixtures to use in a model (so 38 would represent a whole season).}

\item{number_simulations}{This is the number of simulations to use for each game in the remaining season. Defaults to 50,000.}

\item{value_seed}{This is the seed to use for the random numbers in the simulation. Defaults to 120519 (which, IMHO, was a great footballing day).}
}
\description{
This function takes the relevant parameters from the model and assigns them to the approapriate teams in each remaining game of the season.
}
\examples{
run_simulations(
  results_thisSeason = example_thisSeason,
  number_seasons = 1L,
  lookback_rounds = 78L,
  number_simulations = 25000L,
  value_seed = 120519L
  )
}
\keyword{Sport}
